nba-props/requirements.txt
# NBA Props Prediction System - Python Dependencies

# Core dependencies
nba_api==1.4.1              # Unofficial NBA stats API
requests==2.31.0            # HTTP library for Odds API
python-dotenv==1.0.1        # Environment variable management

# Database
SQLAlchemy==2.0.35          # ORM and database toolkit

# Data processing
pandas==2.1.4               # Data manipulation and analysis
numpy==1.26.3               # Numerical computing

# Utilities
pytz==2024.1                # Timezone support
